requirements.txt
langchain
fastapi
uvicorn
langgraph
langchain-ollama
langchain-community
tavily-python
langchain-google-vertexai
google-cloud-aiplatform
langchain-google-vertexai
python-dotenv
langchain-tavily
langchain-google-genai
motor
pymongo
passlib
python-jose
pandas
python-dateutil
pytest
httpx
dotenv
requests
